+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Кекстаграм (простой)»
 
 * Студент: [Никита Тропин](https://up.htmlacademy.ru/univer-js1/3/user/2345999).
-* Наставник: `Неизвестно`.
+* Наставник: [Глеб Клецков](https://htmlacademy.ru/profile/id1487865)
 
 ---
